Split out utility module
It contains pack_timestamp, split_timestamp, pack_query_timestamp (for
now). There were artificially bundled inside Backend, but will also be
needed by an async backend.
